+++ src/pages/commits/ActiveCommits/CommitBarChart/CommitChartTooltip.tsx
@@ -6,10 +6,10 @@ import { Disclaimer } from "@leafygreen-ui/typography";
 import { inactiveElementStyle } from "components/styles";
 import { taskStatusToCopy, mapTaskToBarchartColor } from "constants/task";
 import { size, zIndex } from "constants/tokens";
+import { getStatusesWithZeroCount } from "pages/commits/ActiveCommits/utils";
 import { ColorCount } from "pages/commits/types";
 import { TaskStatus } from "types/task";
 import { msToDuration } from "utils/string";
-import { getStatusesWithZeroCount } from "./utils";
 
 const { gray } = palette;
 interface Props {
