+++ src/pages/commits/ActiveCommits/CommitBarChart/index.tsx
@@ -2,8 +2,8 @@ import styled from "@emotion/styled";
 import { commitChartHeight } from "pages/commits/constants";
 import { ColorCount } from "pages/commits/types";
 import { ChartTypes } from "types/commits";
-import { CommitChartTooltip } from "../CommitChartTooltip";
 import { calculateBarHeight } from "../utils";
+import { CommitChartTooltip } from "./CommitChartTooltip";
 
 interface Props {
   groupedTaskStats: ColorCount[];
@@ -13,7 +13,7 @@ interface Props {
   eta?: Date;
 }
 
-export const CommitChart: React.VFC<Props> = ({
+export const CommitBarChart: React.VFC<Props> = ({
   max,
   chartType,
   groupedTaskStats,
